tui: split minimal and extra themes into welcome, tabs, and modals sub-files
